move mock portal users and banks into own directory
