+++ utils/mock-data-loader/bss/index.js
@@ -1,13 +1,9 @@
-const USERS = require('./users');
-const BANKS = require('./banks');
 const MANDATORY_CRITERIA = require('./mandatoryCriteria');
 const ELIGIBILITY_CRITERIA = require('./eligibilityCriteria');
 const DEALS = require('./deals');
 const FACILITIES = require('./facilities');
 
 const MOCKS = {
-  USERS,
-  BANKS,
   ELIGIBILITY_CRITERIA,
   MANDATORY_CRITERIA,
   DEALS,
