+++ utils/mock-data-loader/portal/index.js
@@ -1,0 +1,7 @@
+const USERS = require('./users');
+
+const MOCKS = {
+  USERS,
+};
+
+module.exports = MOCKS;
